+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Fixed
+- Fix unreachable code panic that happened if the `Receiver` of an empty but open channel was
+  polled and then dropped.
 
 
 ## [0.1.1] - 2020-05-10